Move command implementation outside processor
