+++ behavior-ms/behavior-bruxo/behavior/processing/impl/CMakeLists.txt
@@ -1,0 +1,9 @@
+robocin_cpp_library(
+    NAME impl
+    HDRS impl.h
+    SRCS impl.cpp
+    DEPS common::utility
+         common::output
+         common::geometry
+         protocols::protobufs
+)
